updated settings view
